refactor: Isola constantes e áudio em módulos de utilidades
- Cria a estrutura `src/utils`
- O  arquivo `constantes.py` centraliza todas as configurações globais
- O arquivo `audio.py` isola toda a lógica de carregamento de som.
