+++ assets/imagens/sprites/FantasiaVampiro 19x25.png
